+++ pubspec.yaml
@@ -64,8 +64,8 @@ dev_dependencies:
   msix: ^3.6.1
   pedantic_sensuikan1973: ^3.2.0
   plugin_platform_interface: ^2.1.2
-  url_launcher_platform_interface: ^2.0.5
   shared_preferences_platform_interface: ^2.0.0
+  url_launcher_platform_interface: ^2.0.5
 
 flutter:
   uses-material-design: true
