+++ test/arrange_discs_mode_test.dart
@@ -12,9 +12,9 @@ import '../test_helper/board_finder.dart';
 import '../test_helper/edax_server.dart';
 import '../test_helper/secure_bookmark_mock.dart';
 import 'widget_test_helper/fake_file_selector.dart';
+import 'widget_test_helper/fake_shared_preferences.dart';
 import 'widget_test_helper/fake_url_launcher.dart';
 import 'widget_test_helper/libedax_assets.dart';
-import 'widget_test_helper/shared_preferences_mock.dart';
 
 Future<void> main() async {
   setUpAll(() async {
