+++ test/launch_app_with_empty_file_path_test.dart
@@ -9,8 +9,8 @@ import 'package:pedax/home/setting_dialogs/book_file_path_setting_dialog.dart';
 import '../test_helper/board_finder.dart';
 import '../test_helper/edax_server.dart';
 import '../test_helper/secure_bookmark_mock.dart';
+import 'widget_test_helper/fake_shared_preferences.dart';
 import 'widget_test_helper/libedax_assets.dart';
-import 'widget_test_helper/shared_preferences_mock.dart';
 
 Future<void> main() async {
   setUpAll(() async {
